Adicionado funcionalidades no código v3
