+++ modulo3/projeto-astromatch/astromatch/src/pages/Body/MatchPage/MatchPage.js
@@ -40,7 +40,6 @@ const MatchPage = (props) => {
       <Container>
         {mapMatches}
       </Container>
-      <button onClick={props.clear}>Limpar</button>
     </>
   )
 }
